03-foundry-sdk/agents-service/requirements.txt
python-dotenv==1.2.1
azure-identity==1.26.0b1
azure-ai-projects==2.0.0b2
openai>=1.63.0
requests==2.32.5
